Initial phone UI (#101)
When Wolvic is running on a phone, provide a simple UI if the VR glasses are not attached yet.

The first time that the app is ran, it will display the Terms of Service and the Privacy Policy.
After that, it will open a Web view with an introduction to Wolvic.

Main changes in VR code:
* LegalDocumentView, LegalDocumentDialogWidget: contain the shared code for displaying the
Terms of Service and the Privacy Policy on startup
* SettingsStore: stores and retrieves whether the Terms of Service have been seen by the user
* SettingsView, SettingsWidget: add an item in the Settings to display the Terms of Service

Main changes in phone UI code:
* PlatformActivity: sets the phone-specific theme and content layout
* ContentHolderFragment: handles the initial content logic in the phone UI; embeds
PrivacyPolicyFragment, * TermsServiceFragment , or LandingPageFragment
* PrivacyPolicyFragment, TermsServiceFragment: only shown on startup, contain buttons
 to accept or reject each legal document
* LandingPageFragment: holds a Web view, a menu to review the legal documents, and
 a button to display the "Enter VR" message
* LegalDocumentFragment: used to display the legal documents from the main page
of the app, include a back arrow to navigate back
* EnterVrFragment: advises the user to put on the VR glasses to use the app
* themes.xml: declares the theme Theme.WolvicPhone that is used by the phone UI;
we need to change the theme because the one used by the VR interface works badly
 with default Android widgets